# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant API",
        "description": "Dokumentasi API autentikasi untuk aplikasi restoran.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/orders/manual": {
            "post": {
                "tags": [
                    "Manual Orders"
                ],
                "summary": "Buat order manual by admin",
                "operationId": "f3eb33bc0fda35f5f1ccf323bc205941",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_type",
                                    "customer_name",
                                    "items",
                                    "payment"
                                ],
                                "properties": {
                                    "order_type": {
                                        "type": "string",
                                        "enum": [
                                            "dine-in",
                                            "take-away"
                                        ],
                                        "example": "dine-in"
                                    },
                                    "table_id": {
                                        "type": "string",
                                        "example": "TB001"
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "customer_phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "string"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "variant": {
                                                    "type": "string",
                                                    "enum": [
                                                        "Ice",
                                                        "Hot"
                                                    ]
                                                },
                                                "unit_price": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "payment": {
                                        "properties": {
                                            "method": {
                                                "type": "string",
                                                "enum": [
                                                    "cash",
                                                    "qris"
                                                ]
                                            },
                                            "amount": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "order_id": {
                                                    "type": "string"
                                                },
                                                "payment_id": {
                                                    "type": "string"
                                                },
                                                "total": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Gagal membuat order"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/manual/available-tables": {
            "get": {
                "tags": [
                    "Manual Orders"
                ],
                "summary": "Dapatkan daftar meja yang tersedia",
                "operationId": "6989f41dfcf877057ffb22f36ebb0f56",
                "responses": {
                    "200": {
                        "description": "List meja tersedia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "table_number": {
                                                        "type": "string"
                                                    },
                                                    "capacity": {
                                                        "type": "integer"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}